Delete to Cart according to user id

+++ Database/Cart.php
@@ -51,7 +51,7 @@ class Cart {
     // delete cart item using cart item id
     public function deleteCart($item_id = null, $table="cart"){
         if($item_id != null){
-            $result = $this->db->con->query("DELETE FROM {$table} WHERE item_id={$item_id}");
+            $result = $this->db->con->query("DELETE FROM {$table} WHERE user_id={$_SESSION['id']} and item_id={$item_id}");
         }
          // if you get a result, refresh the browser
         if($result){
